web/requirements.txt
aioredis==1.2.0
amqp==2.3.2
asgi-redis==1.0.0
asgiref==2.3.0
astroid==2.1.0
async-timeout==2.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.0
certifi==2018.11.29
channels==2.1.6
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
daphne==2.2.4
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
ebaysdk==2.1.5
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.4
jsonfield==2.0.2
kombu==4.2.2.post1
lazy-object-proxy==1.3.1
lxml==4.3.0
mccabe==0.6.1
msgpack==0.6.1
msgpack-python==0.5.6
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.9
redis==2.10.6
requests==2.21.0
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
vine==1.2.0
wrapt==1.10.11
zope.interface==4.6.0
psycopg2>=2.7,<3.0
gunicorn==19.6.0
psycopg2-binary
